Add VFD with VirtualBox Guest Addition modules
So that when you are using Grml on a VirtualBox VM, you don't have to
recompile the modules every time but can just insert the VFD into its
virtual floppy drive and load the modules.

Note that for grml96 you have to run this script twice (once for 32-bit
and once for 64-bit) if you want to have both VFDs.