Add USB loader.
This provides an ISO and a floppy image that can be used to boot grml-plus
from CD/floppy. It includes multiple approaches

- PLOP Boot Manager
- Use GRUB's USB support to chainload a multiboot GRUB image from the USB
  key, which will then load the grub.cfg from USB. The additional image is
  used so that the CD does not need to be updated if the new grml version
  includes a different GRUB version with incompatible modules.
- Chainload a standalone multiboot GRUB image that chainloads the GRUB
  image mentioned above. This way you don't have to rely on the (quite
  unstable) nativedisk switch support of GRUB, and the second stage image
  can easily be added to a ISOLINUX/GRUB4DOS based boot disk.

The image also includes Kon-Boot, Memtest and iPXE, as these are quite
small (and rarely updated) and probably useful to load without
chainloading from USB. The CD image also includes a copy of the all-in-one
boot floppy.