Add Windows 10 Recovery
To use, run recoverydrive.exe from within Windows 10 (tested with the 1511
build), then pass the path to the generated mounted drive as argument to
add-windows-10-recovery.

For 64-bit Windows 10, booting both EFI and BIOS works. Files are stored in
/sources, /boot and /efi/microsoft - to make the files less cluttered, edit
the BCD with bcdedit.exe to move the boot.wim from /sources to /boot.